positions, history, and algorithm-list to other files.
